시도 전

+++ server/movies/models.py
@@ -66,7 +66,7 @@ class MovieReview(models.Model):
     user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
     movies = models.ForeignKey(Movie, on_delete=models.CASCADE)
     content = models.TextField()
-    vote = models.IntegerField()
+    vote = models.FloatField()
     block_users = models.ManyToManyField(settings.AUTH_USER_MODEL, related_name='block_reviews')
     banned = models.BooleanField(default=False)
     created_at = models.DateTimeField(auto_now_add=True)
